--- .vscode/tasks.json
@@ -1,45 +1,0 @@
-{
-	"version": "2.0.0",
-	"tasks": [
-		{
-			"label": "dev",
-			"type": "npm",
-			"script": "dev",
-			"group": "build",
-			"isBackground": true,
-			"problemMatcher": {
-				"owner": "vite",
-				"pattern": {
-					"regexp": ".*"
-				},
-				"background": {
-					"activeOnStart": true,
-					"beginsPattern": "Local:",
-					"endsPattern": "ready in"
-				}
-			}
-		},
-		{
-			"label": "dev-sod",
-			"type": "shell",
-			"command": "node",
-			"args": ["scripts/devSod.js"],
-			"group": "build",
-			"isBackground": true,
-			"options": {
-				"cwd": "${workspaceFolder}"
-			},
-			"problemMatcher": {
-				"owner": "vite",
-				"pattern": {
-					"regexp": ".*"
-				},
-				"background": {
-					"activeOnStart": true,
-					"beginsPattern": "Local:",
-					"endsPattern": "ready in"
-				}
-			}
-		}
-	]
-}
